# Fix a scaling issue in MagneticField/VolumeBasedEngine