@cmsbuild 2019-10-15
-code-checks

Logs: https://cmssdt.cern.ch/SDT/code-checks/cms-sw-PR-28180/12257

- This PR adds an extra 20KB to repository

Code check has found code style and quality issues which could be resolved by applying following patch(s)
- **code-format**:
https://cmssdt.cern.ch/SDT/code-checks/cms-sw-PR-28180/12257/code-format.patch
e.g. `curl https://cmssdt.cern.ch/SDT/code-checks/cms-sw-PR-28180/12257/code-format.patch | patch -p1`
You can also run `scram build code-format` to apply code format directly