@civanch 2019-10-15
@amadio , this is a very good fix. You need to fix code-format issues and may be remove "cout", which is left in the code.