@cmsbuild 2019-10-16
A new Pull Request was created by  @amadio (Guilherme Amadio) for master.

It involves the following packages:

MagneticField/GeomBuilder
MagneticField/VolumeBasedEngine


@perrotta, @cmsbuild, @slava77 can you please review it and eventually sign? Thanks.
@namapane this is something you requested to watch as well.
@davidlange6, @slava77, @fabiocos you are the release manager for this.

cms-bot commands are listed <a href="http://cms-sw.github.io/cms-bot-cmssw-cmds.html">here</a>