@namapane 2019-10-16
> 
> 
> @civanch, Thank you. I wondered if it was ok to remove, since it can be a performance issue, and it's easy to add back if needed, but decided to make only the minimal changes. I will fix formatting and make other improvements like removing the print out calls.

@amadio, great job. Regarding the couts: they occur only in case of errors so should not be a real performance issue, and would better be replaced with LogError than removed.