@amadio 2019-10-18
As for the printouts, I can see a small performance increase when I remove them. The branch is still there at runtime even when the `if` is almost always `false`. For a function called millions of times like this, it might make sense to remove or use `#ifdef` to not pay this cost.