@namapane 2019-10-18
> 
> 
> As for the printouts, I can see a small performance increase when I remove them. The branch is still there at runtime even when the `if` is almost always `false`. For a function called millions of times like this, it might make sense to remove or use `#ifdef` to not pay this cost.

Which couts exactly are you referring to in this case (which line numbers)?

PS: as for the previous comment, I was actually argumenting **against** letting the caller handle the caches.