@civanch 2019-10-18
@amadio , for SIM we do not need static thread local cache, because Field object is unique per thread, so any cache value/pointer are simply class members. I am personally struggle against overuse of thread local storage in Geant4 and we slowly reduce it size. I do not know about RECO field, is it really shared between threads?