@amadio 2019-10-18
@civanch Can we then make the field in this PR also a unique instance for each thread? We may have some hurdles (like the vectors taken by reference at construction, then deleting their members in the destructor, which would prevent reusing the vectors for all threads), but it might be a good way to avoid the performance issue with `std::atomic` without needing to add a `thread_local` volume cache.