@Dr15Jones 2019-10-18
>  I do not know about RECO field, is it really shared between threads?

Yes, all EventSetup objects are shared by all threads.