@civanch 2019-10-18
@Dr15Jones , normally we should share static data or static objects with const methods, concrete field class, which access such static objects in case of SIM is unique per thread and cache may be implement there.