@Dr15Jones 2019-10-18
One way to approach this would be to to add a new interface to the MagneticField class which allows for external caching (this can be done in a type agnostic way). Then have a 'local' MagneticField class which is a wrapper around the shared MagneticField and the cache which is passed into the shared MagneticField.