@Dr15Jones 2019-10-18
@civanch 
Shared non-const static objects are against the CMSSW rules as they are inherently non-thread safe (plus other problems with long term maintenance).