@civanch 2019-10-18
Sure, I agree.