@namapane 2019-10-18
> 
> 
> One way to approach this would be to to add a new interface to the MagneticField class which allows for external caching (this can be done in a type agnostic way). Then have a 'local' MagneticField class which is a wrapper around the shared MagneticField and the cache which is passed into the shared MagneticField.

Uhm, I have a hard time figuring out what you mean here. Can you provide a more concrete example? 
My point is: in the reco code, the user gets  a pointer to a MF "engine" which in the most general case does not use volumes at all. There's just one specific type of engine which internally uses volumes, it's not the only one. The user code should not normally know whether caching is necessary or not. It is not clear to me how you would get around this?