@namapane 2019-10-18
> 
> 
> @civanch Can we then make the field in this PR also a unique instance for each thread? We may have some hurdles (like the vectors taken by reference at construction, then deleting their members in the destructor, which would prevent reusing the vectors for all threads), but it might be a good way to avoid the performance issue with `std::atomic` without needing to add a `thread_local` volume cache.



> 
> 
> > I do not know about RECO field, is it really shared between threads?
> 
> Yes, all EventSetup objects are shared by all threads.

Indeed; and besides, a full volume-based map takes some memory (something like ~60 MB of raw data alone) which would be a waste to replicate in each thread, even if that was possible.