@civanch 2019-10-18
We all agree that there is a static data as a field and static object in event setup to access this data. 

I like a proposal to change the interface allowing to keep cache in external classes, which calls this static object. For me, it is not a good solution if a static object has cache implemented as thread_local or atomic. In SIM we already instantiate field class per thread following Geant4 interface. The interface to the CMS field may be changed to be maximally effective. It is a very local place in SIM, we can do whatever reasonable modification.