@namapane 2019-10-18
@civanch I assume you mean that in SIM the G4 field interface is instantiated per-thread, but that is just a wrapper of the actual field engine that is common to all threads (otherwise we would not be discussing this issue, which was observed in SIM if I understand correctly).
This said I do not understand what is the proposal. In RECO the usage of the field is ubiquitous, we can sure make changes to accomodate the need for SIM but I don't think we can substantially change the interface for RECO. (But maybe I'm missing something on what you have in mind)