@Dr15Jones 2019-10-18
@civanch 
>We all agree that there is a static data as a field and static object in event setup to access this data.

I assume by static you do not mean C++ `static` (which I completetly disagree with) but with _immutable_, i.e. unchanging (which I do agree with).