@civanch 2019-10-18
@Dr15Jones , yes immutable object with field data.
@namapane , I mean, that ideally cache should be in wrapper object and not in field engine.